Move voice endpoints into voice package and restrict scope.
