+++ src/main/java/com/nexmo/client/voice/CreateCallMethod.java
@@ -19,38 +19,42 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice.endpoints;
-
+package com.nexmo.client.voice;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
-import com.nexmo.client.NexmoUnexpectedException;
 import com.nexmo.client.auth.JWTAuthMethod;
-import com.nexmo.client.voice.CallInfoPage;
-import com.nexmo.client.voice.CallsFilter;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
 import org.apache.http.client.methods.RequestBuilder;
-import org.apache.http.client.utils.URIBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.BasicResponseHandler;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.net.URISyntaxException;
-import java.util.List;
 
-public class ListCallsMethod extends AbstractMethod<CallsFilter, CallInfoPage> {
+class CreateCallMethod extends AbstractMethod<Call, CallEvent> {
     private static final Log LOG = LogFactory.getLog(CreateCallMethod.class);
 
     private static final String PATH = "/calls";
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{JWTAuthMethod.class};
     private String uri;
 
-    public ListCallsMethod(HttpWrapper httpWrapper) {
+    CreateCallMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
+    }
+
+    @Override
+    public RequestBuilder makeRequest(Call request) throws NexmoClientException, UnsupportedEncodingException {
+        // TODO: Remove in 4.0.0 along with setUri method
+        String uri = (this.uri != null) ? this.uri : httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH;
+
+        return RequestBuilder
+                .post(uri)
+                .setHeader("Content-Type", "application/json")
+                .setEntity(new StringEntity(request.toJson()));
     }
 
     @Override
@@ -59,29 +63,9 @@ public class ListCallsMethod extends AbstractMethod<CallsFilter, CallInfoPage> {
     }
 
     @Override
-    public RequestBuilder makeRequest(CallsFilter filter) throws NexmoClientException, UnsupportedEncodingException {
-        URIBuilder uriBuilder;
-        // TODO: Remove in 4.0.0 along with setUri and getUri method
-        String uri = (this.uri != null) ? this.uri : httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH;
-
-        try {
-            uriBuilder = new URIBuilder(uri);
-        } catch (URISyntaxException e) {
-            throw new NexmoUnexpectedException("Could not parse URI: " + uri);
-        }
-        if (filter != null) {
-            List<NameValuePair> params = filter.toUrlParams();
-            for (NameValuePair param : params) {
-                uriBuilder.setParameter(param.getName(), param.getValue());
-            }
-        }
-        return RequestBuilder.get().setUri(uriBuilder.toString());
-    }
-
-    @Override
-    public CallInfoPage parseResponse(HttpResponse response) throws IOException {
+    public CallEvent parseResponse(HttpResponse response) throws IOException {
         String json = new BasicResponseHandler().handleResponse(response);
-        return CallInfoPage.fromJson(json);
+        return CallEvent.fromJson(json);
     }
 
     /**
@@ -90,9 +74,5 @@ public class ListCallsMethod extends AbstractMethod<CallsFilter, CallInfoPage> {
     @Deprecated
     public void setUri(String uri) {
         this.uri = uri;
-    }
-
-    public String getUri() {
-        return (this.uri != null) ? this.uri : httpWrapper.getHttpConfig().getApiBaseUri() + PATH;
     }
 }
