+++ src/main/java/com/nexmo/client/voice/StartTalkMethod.java
@@ -19,33 +19,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice.endpoints;
+package com.nexmo.client.voice;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.JWTAuthMethod;
-import com.nexmo.client.voice.Recording;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.BasicResponseHandler;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-public class DownloadRecordingEndpoint extends AbstractMethod<String, Recording> {
-    private static final Log LOG = LogFactory.getLog(DownloadRecordingEndpoint.class);
+class StartTalkMethod extends AbstractMethod<TalkRequest, TalkResponse> {
+    private static final Log LOG = LogFactory.getLog(StartTalkMethod.class);
 
+    private static final String PATH = "/calls/";
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{JWTAuthMethod.class};
+    private String uri;
 
-    public DownloadRecordingEndpoint(HttpWrapper httpWrapper) {
+    StartTalkMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
-    }
-
-    @Override
-    public RequestBuilder makeRequest(String uri) throws NexmoClientException, UnsupportedEncodingException {
-        return RequestBuilder.get().setUri(uri);
     }
 
     @Override
@@ -54,7 +52,28 @@ public class DownloadRecordingEndpoint extends AbstractMethod<String, Recording>
     }
 
     @Override
-    public Recording parseResponse(HttpResponse response) throws IOException {
-        return new Recording(response);
+    public RequestBuilder makeRequest(TalkRequest request) throws NexmoClientException, UnsupportedEncodingException {
+        // TODO: Remove in 4.0.0 along with setUri method
+        String baseUri = (this.uri != null)
+                ? this.uri
+                : httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH;
+        return RequestBuilder
+                .put(baseUri + request.getUuid() + "/talk")
+                .setHeader("Content-Type", "application/json")
+                .setEntity(new StringEntity(request.toJson()));
+    }
+
+    @Override
+    public TalkResponse parseResponse(HttpResponse response) throws IOException {
+        String json = new BasicResponseHandler().handleResponse(response);
+        return TalkResponse.fromJson(json);
+    }
+
+    /**
+     * @deprecated Use {@link com.nexmo.client.HttpConfig.Builder} to create an {@link com.nexmo.client.HttpConfig} object and pass into {@link com.nexmo.client.NexmoClient}
+     */
+    @Deprecated
+    public void setUri(String uri) {
+        this.uri = uri;
     }
 }
