+++ src/main/java/com/nexmo/client/voice/StopTalkMethod.java
@@ -19,44 +19,30 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.voice.endpoints;
+package com.nexmo.client.voice;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.JWTAuthMethod;
-import com.nexmo.client.voice.Call;
-import com.nexmo.client.voice.CallEvent;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
-import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.BasicResponseHandler;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-public class CreateCallMethod extends AbstractMethod<Call, CallEvent> {
-    private static final Log LOG = LogFactory.getLog(CreateCallMethod.class);
+class StopTalkMethod extends AbstractMethod<String, TalkResponse> {
+    private static final Log LOG = LogFactory.getLog(StopTalkMethod.class);
 
-    private static final String PATH = "/calls";
+    private static final String DEFAULT_URI = "https://api.nexmo.com/v1/calls/";
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{JWTAuthMethod.class};
-    private String uri;
+    private String uri = DEFAULT_URI;
 
-    public CreateCallMethod(HttpWrapper httpWrapper) {
+    StopTalkMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
-    }
-
-    @Override
-    public RequestBuilder makeRequest(Call request) throws NexmoClientException, UnsupportedEncodingException {
-        // TODO: Remove in 4.0.0 along with setUri method
-        String uri = (this.uri != null) ? this.uri : httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH;
-
-        return RequestBuilder
-                .post(uri)
-                .setHeader("Content-Type", "application/json")
-                .setEntity(new StringEntity(request.toJson()));
     }
 
     @Override
@@ -65,15 +51,17 @@ public class CreateCallMethod extends AbstractMethod<Call, CallEvent> {
     }
 
     @Override
-    public CallEvent parseResponse(HttpResponse response) throws IOException {
-        String json = new BasicResponseHandler().handleResponse(response);
-        return CallEvent.fromJson(json);
+    public RequestBuilder makeRequest(String uuid) throws NexmoClientException, UnsupportedEncodingException {
+        String uri = this.uri + uuid + "/talk";
+        return RequestBuilder.delete(uri).setHeader("Content-Type", "application/json");
     }
 
-    /**
-     * @deprecated Use {@link com.nexmo.client.HttpConfig.Builder} to create an {@link com.nexmo.client.HttpConfig} object and pass into {@link com.nexmo.client.NexmoClient}
-     */
-    @Deprecated
+    @Override
+    public TalkResponse parseResponse(HttpResponse response) throws IOException {
+        String json = new BasicResponseHandler().handleResponse(response);
+        return TalkResponse.fromJson(json);
+    }
+
     public void setUri(String uri) {
         this.uri = uri;
     }
