+++ src/main/java/com/nexmo/client/voice/TalkEndpoint.java
@@ -19,8 +19,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package com.nexmo.client.voice;
 
-/**
- * Provides internal classes and interfaces implementing the Nexmo Voice API.
- */
-package com.nexmo.client.voice.endpoints;
+import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.NexmoClientException;
+
+import java.io.IOException;
+
+public class TalkEndpoint {
+    private final StartTalkMethod startTalk;
+    private final StopTalkMethod stopTalk;
+
+    public TalkEndpoint(HttpWrapper wrapper) {
+        this.startTalk = new StartTalkMethod(wrapper);
+        this.stopTalk = new StopTalkMethod(wrapper);
+    }
+
+    public TalkResponse put(TalkRequest request) throws IOException, NexmoClientException {
+        return this.startTalk.execute(request);
+    }
+
+    public TalkResponse delete(String uuid) throws IOException, NexmoClientException {
+        return this.stopTalk.execute(uuid);
+    }
+}
